chore: refact code

+++ src/core-haskell/resolvers/Helpers.hs
@@ -36,7 +36,7 @@ isRevenue t = (value t) >= 0 && (isRevenueOrExpense t)
 isExpense :: Transaction -> Bool
 isExpense t = (value t) < 0 && (isRevenueOrExpense t)
 
-mean :: [Float] -> Float
+mean :: [Double] -> Double
 mean values = sum (values) / fromIntegral (length values)
 
 -- Agrupa a lista de transações por dia.
@@ -48,24 +48,24 @@ groupTransactionsByMonth :: [Transaction] -> [[Transaction]]
 groupTransactionsByMonth transactions = (groupBy ((==) `on` getMonth) transactions)
 
 -- Retorna uma lista contendo a sobra para cada dia da lista de transações.
-getDaysRemains :: [Transaction] -> [Float]
+getDaysRemains :: [Transaction] -> [Double]
 getDaysRemains transactions = map (getRemains) (groupTransactionsByDay transactions)
 
 -- Retorna uma lista contendo a sobra para cada dia da lista de transações.
-getMonthsRemains :: [Transaction] -> [Float]
+getMonthsRemains :: [Transaction] -> [Double]
 getMonthsRemains transactions = map (getRemains) (groupTransactionsByMonth transactions)
 
 -- Retorna a sobra para a lista de transações.
-getRemains :: [Transaction] -> Float
+getRemains :: [Transaction] -> Double
 getRemains transactions = sum (map (value) transactions)
 
 -- Retorna uma lista contendo o saldo para cada dia da lista de transações.
-getDaysBalances :: [Transaction] -> Float -> [Float]
+getDaysBalances :: [Transaction] -> Double -> [Double]
 getDaysBalances revenuesAndExpenses initialBalance = 
   snd (mapAccumL (\x y -> (x+y,x+y)) initialBalance (getDaysRemains revenuesAndExpenses))
 
 -- Retorna o saldo inicial de uma lista de transaçãoes.
-getInitialBalance :: [Transaction] -> Float
+getInitialBalance :: [Transaction] -> Double
 getInitialBalance transactions = value (transactions !! 0)
 
 -- Retorna o dia do mês de uma transação.
